School exam PHP
The webpage I created for school project. In that time I had no idea on sql or any other things is that it was not required to create and link any database too.  My teacher only asked to show three time zone clock on the pages.  PHP is server side so end user can't see the right time until the page is refreshed. I was stubborn and found ajax and little Javascript to refresh the page every second.